+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Changed
 
 * Require Emacs 26.1 (released in 2018) instead of 25.1.
+* When eide-custom-smooth-scrolling option is enabled, increase
+  mouse-wheel-scroll-amount to 3, otherwise the scrolling is too slow.
 
 ### Fixed